implementazione Curatore

+++ src/main/java/unicam/progettofiliera/infrastructure/ApprovazioneProdottiRepository.java
@@ -3,5 +3,5 @@ package unicam.progettofiliera.infrastructure;
 import org.springframework.data.jpa.repository.JpaRepository;
 import unicam.progettofiliera.models.prodotti.Prodotto;
 
-public interface ApprovazioneProdottiRepository extends JpaRepository<Prodotto, Integer> {
+public interface ApprovazioneProdottiRepository extends JpaRepository<Prodotto, Long> {
 }
